refactor(ui): extract coordinator-admin/components/groups-panel.ts
Move renderGroupsPanel into its own module. The panel now takes a deps
object with { summary, createGroup, runGroup, renderShell, reloadData }
so the barrel wires in the factory actions, the shared shell re-render,
and the data reload invoked when selecting a group to manage.
